% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.best.templates.R
\name{get.best.templates}
\alias{get.best.templates}
\title{Selects the \emph{best} template for a given signal based on fitting
       values returned by a comparison function}
\usage{
get.best.templates(
  time.instants,
  signal,
  templates,
  referential.time.instant = 0,
  delta.time.before.ref = 0,
  delta.time.after.ref = 20 * 0.8,
  comparison.function = get.MSE,
  keep.details = TRUE,
  time.tol = min(diff(time.instants))/100,
  ...
)
}
\arguments{
\item{time.instants}{The time instants in which the reference signals
and templates are sampled}

\item{signal}{The reference signal to select the best template}

\item{templates}{A list of signals from which the "most similar" to the
reference signal will be chosen}

\item{referential.time.instant}{The time instant to be used as a reference
in the comparison of the signal and the templates
Default value: 0}

\item{delta.time.before.ref}{The amount of time before the referential
instant to be included in the comparison
Default value: 0}

\item{delta.time.after.ref}{The amount of time after the referential
instant to be included in the comparison.
Default value: 20 * 0.8 s (20 heart beats approx.)}

\item{comparison.function}{The function to be used to compare and estimate
the goodness-of-fit between each template and the referential signal
in the period of time of interest.
Default value: get.MSE}

\item{keep.details}{Whether the answer should keep all comparisons details
Default value: TRUE}

\item{time.tol}{The minimum amount of time between two time instants.
Default value: one-hundredth of the minimum time difference
between tow time instants}

\item{...}{Arguments pass over to comparison.function()}
}
\value{
A list with (at least):
\describe{
  \item{best.template.index}{The index of the best template in the giving
        list of templates}
  \item{best.fit.value}{The fitting value scored by the best template}
}
}
\description{
Selects the \emph{best} template for a given signal based on fitting
       values returned by a comparison function
}
