% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.COR.R
\name{get.COR}
\alias{get.COR}
\alias{get.COR.default}
\title{Gets the correlation coefficient (COR) between two signals}
\usage{
get.COR(sim, obs, ...)

\method{get.COR}{default}(
  sim,
  obs,
  use = "everything",
  method = c("pearson", "kendall", "spearman"),
  ...
)
}
\arguments{
\item{sim}{Simulated (estimated) signal}

\item{obs}{Observed (actual) signal}

\item{...}{Arguments pass over to cor()}

\item{use}{Pass to \code{use} parameter of the \code{cor} function}

\item{method}{Pass to \code{method} parameter of the \code{cor} function}
}
\description{
Gets the correlation coefficient (COR) between two signals
}
\section{Methods (by class)}{
\itemize{
\item \code{get.COR(default)}: Default implementation
(for numeric vectors and \code{ts} objects)

}}
