% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.MSE.R
\name{get.MSE}
\alias{get.MSE}
\alias{get.MSE.default}
\alias{get.MSE.matrix}
\alias{get.MSE.data.frame}
\title{Gets the mean squared error (MSE) between two signals}
\usage{
get.MSE(sim, obs, ...)

\method{get.MSE}{default}(sim, obs, ...)

\method{get.MSE}{matrix}(sim, obs, ...)

\method{get.MSE}{data.frame}(sim, obs, ...)
}
\arguments{
\item{sim}{Simulated (estimated) signal}

\item{obs}{Observed (actual) signal}

\item{...}{Arguments pass over to mean()}
}
\description{
Gets the mean squared error (MSE) between two signals
}
\section{Methods (by class)}{
\itemize{
\item \code{get.MSE(default)}: Default implementation
(for vectors and \code{ts} objects)

\item \code{get.MSE(matrix)}: Implementation for matrices

\item \code{get.MSE(data.frame)}: Implementation for data frames

}}
