% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.normalised.ABP.stimulus.R
\name{get.normalised.ABP.stimulus}
\alias{get.normalised.ABP.stimulus}
\title{Creates a normalised step decrease signal to represent an arterial blood
pressure (ABP) stimulus.}
\usage{
get.normalised.ABP.stimulus(
  sampling.time = 0.1,
  time.until.release = 10,
  time.after.release = 20,
  smooth.step.stimulus = FALSE,
  filter.order = 2,
  cutoff.frequency = 0.2,
  left.stabilisation.time = ifelse(smooth.step.stimulus, 30, 0),
  time.rounding.digits = format.info(sampling.time)[2],
  time.tol = sampling.time/100
)
}
\arguments{
\item{sampling.time}{Sampling time for the ABP stimulus (in seconds).
Default value: 0.1 s}

\item{time.until.release}{Time included in the resulting signal until
the time of cuff release (in seconds).
Default value: 10 s}

\item{time.after.release}{Time included in the resulting signal after
the time of cuff release.
Default value: 20 s}

\item{smooth.step.stimulus}{Logical, whether the ABP step stimulus
should be smoothed (filtered).
Default value: FALSE}

\item{filter.order}{The order of the low-pass Butterworth filter used
to smooth the ABP step stimulus.
Default value: 2}

\item{cutoff.frequency}{The cutoff frequency for the low-pass Butterworth
filter used to smooth the ABP step stimulus (in Hz).
Default value: 0.20 Hz}

\item{left.stabilisation.time}{The amount of time to be used for the
stabilisation of the decrease in ABP. This time is included in the
step decrease when the filter is applied, but removed after that.
Default value: 30 s when the filter will be applied, 0 s otherwise}

\item{time.rounding.digits}{The number of decimal places used for
time instants.
Default value: the number of decimals in the sampling time}

\item{time.tol}{The minimum amount of time between two time instants.
       Default value: one-hundredth of the sampling time

 @return A list with the following:
 \describe{
   \item{time.instants}{The time instants in which the signal is sampled}
   \item{ABP.normalised}{The normalised ABP stimulus generated}
   \item{sampling.time}{The sampling time used}
   \item{time.release}{The time instant in which the thigh cuffs are
         released. Currently at the second 0.0}
 }}
}
\description{
This function creates a normalised step decrease signal, with baseline
value one that drops to zero, to be used as a (rather perfect) ABP
stimulus, similar to the ones observed after the release of thigh cuffs.
}
\details{
The stimulus can be 'smoothed' to get 'more realistic' step signals.
}
