% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.ABP.signal.R
\name{normalise.ABP.signal}
\alias{normalise.ABP.signal}
\title{Normalises a signal of arterial blood pressure (ABP)}
\usage{
normalise.ABP.signal(
  time.instants,
  ABP.signal,
  sample.release,
  sampling.time,
  baseline.initial.time = time.instants[1],
  baseline.final.time = time.instants[sample.release],
  min.ABP.max.delta.time = 5 * 0.8,
  time.tol = sampling.time/100
)
}
\arguments{
\item{time.instants}{The time instants in which the ABP signal is sampled}

\item{ABP.signal}{The original ABP signal}

\item{sample.release}{The index of the sample when the thigh cuffs were
realeased}

\item{sampling.time}{The sampling time used in the ABP signal}

\item{baseline.initial.time}{The time instant that marks the beggining
of the signal baseline.
Default value: \code{time.instants[1]}}

\item{baseline.final.time}{The time instant that marks the end
of the signal baseline.
Default value: \code{time.instants[sample.release]}}

\item{min.ABP.max.delta.time}{The amount of time to search for the minimum
ABP value after the realease of the thigh cuffs.
Default value: 5 * 0.8 s (5 heart beats approx.)}

\item{time.tol}{The minimum amount of time between two time instants.
       Default value: one-hundredth of the sampling time

 @return A list with the following:
 \describe{
   \item{<*details*>}{to be able to normalised the signal}
   \item{normalised.ABP.signal}{The normalised ABP signal}
 }

When the stimulus used to assess the dynamic cerebral autoregulation is not
based on thigh cuffs, \code{sample.release} should indicate the index of the
sample in which the ABP signal should start the expected drop.
It is important to identified this point in the signal to estimate the
signal's baseline value and locate its minimum value.}
}
\description{
Normalises a signal of arterial blood pressure (ABP)
}
