% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.CBFV.signal.R
\name{normalise.CBFV.signal}
\alias{normalise.CBFV.signal}
\title{Normalises a signal of cerebral blood flow velocity (CBFV)}
\usage{
normalise.CBFV.signal(
  time.instants,
  CBFV.signal,
  sample.release,
  sampling.time,
  baseline.initial.time = time.instants[1],
  baseline.final.time = time.instants[sample.release],
  min.CBFV.max.delta.time = 8 * 0.8,
  time.tol = sampling.time/100
)
}
\arguments{
\item{time.instants}{The time instants in which the ABP signal
is sampled}

\item{CBFV.signal}{The original CBFV signal}

\item{sample.release}{The index of the sample when thigh cuffs were
realeased}

\item{sampling.time}{The sampling time used in the CBFV signal}

\item{baseline.initial.time}{The time instant that marks the beggining
of the signal baseline.
Default value: \code{time.instants[1]}}

\item{baseline.final.time}{The time instant that marks the end
of the signal baseline.
Default value: \code{time.instants[sample.release]}}

\item{min.CBFV.max.delta.time}{The amount of time to search for the
minimum CBFV value after the realease of the thigh cuffs.
Default value: 8 * 0.8 s (8 heart beats approx.)}

\item{time.tol}{The minimum amount of time between two time instants.
       Default value: \code{time.instants[sampling.time]} / 100

 @return A list with the following:
 \describe{
   \item{<*details*>}{to be able to normalised the signal}
   \item{normalised.ABP.signal}{The normalised ABP signal}
 }}
}
\description{
Normalises a signal of cerebral blood flow velocity (CBFV)
}
